Version 1.4.1.
Improves compression ratio and compression speed for level 1 for data which contains sequences of the same byte (such as zero initialized data or same-colored pixels in raw images).

An example is a byte sequence like 0x85, 0x1b, 0xfa, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xa0, 0xc8, 0xc2 where version 1.4.0 in level 1 would require an additional sequence of 0x00 in the past in order to compress it. This restriction has been removed in 1.4.1.